--- plugin.yml
@@ -1,4 +1,0 @@
-name: JoinAnnouncements
-main: me.noodles.join.MainJoin
-version: 0.2
-author: BGHDNetwork
